Adjust project structure and imports to clearly separate DB concerns
