+++ app/database/models.py
@@ -2,7 +2,7 @@ from sqlalchemy import create_engine, Column, Integer, String, Float, DateTime, 
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 from datetime import datetime
-from app.database_base import Base
+from app.database.base import Base
 from app.repositories.drift_repository import DriftChangePoint
 import os
